+++ src/bolig_ping/email.py
@@ -1,35 +1,35 @@
-"""Sending emails with flat listings."""
+"""Sending emails with home listings."""
 
 import yagmail
 from dotenv import load_dotenv
 
-from .data_models import Flat
+from .data_models import Home
 
 load_dotenv()
 
 
-def compose_email(flats: list[Flat]) -> tuple[str, str]:
-    """Compose an email with the given flats.
+def compose_email(homes: list[Home]) -> tuple[str, str]:
+    """Compose an email with the given homes.
 
     Args:
-        flats:
-            The flats to compose the email with.
+        homes:
+            The homes to compose the email with.
 
     Returns:
         A pair (subject, contents) for the email.
     """
-    match len(flats):
+    match len(homes):
         case 0:
-            raise ValueError("Cannot compose an email with no flats.")
+            raise ValueError("Cannot compose an email with no homes.")
         case 1:
-            subject = "[BoligPing] Found a new flat!"
-            contents = "Hi,\n\nI found a new flat that you might be interested in:\n\n"
+            subject = "[BoligPing] Found a new home!"
+            contents = "Hi,\n\nI found a new home that you might be interested in:\n\n"
         case _:
-            subject = f"[BoligPing] Found {len(flats)} new flats!"
+            subject = f"[BoligPing] Found {len(homes)} new homes!"
             contents = (
-                "Hi,\n\nI found some new flats that you might be interested in:\n\n"
+                "Hi,\n\nI found some new homes that you might be interested in:\n\n"
             )
-    contents += "\n\n".join(flat.to_html() for flat in flats)
+    contents += "\n\n".join(home.to_html() for home in homes)
     contents += "\n\nHave a splendid day!\n\nBest regards,\nBoligPing"
     return subject, contents