+++ tests/test_cache.py
@@ -6,16 +6,16 @@ from pathlib import Path
 import pytest
 
 from bolig_ping.cache import store_to_cache
-from bolig_ping.data_models import Flat
+from bolig_ping.data_models import Home
 
 
 class TestStoreToCache:
     """Tests for the store_to_cache function."""
 
     @pytest.fixture(scope="class")
-    def flat(self) -> Generator[Flat, None, None]:
-        """Return a Flat object."""
-        yield Flat(
+    def home(self) -> Generator[Home, None, None]:
+        """Return a Home object."""
+        yield Home(
             url="https://some.url",
             address="Some address",
             price=1000,
@@ -26,9 +26,9 @@ class TestStoreToCache:
         )
 
     @pytest.fixture(scope="class")
-    def another_flat(self) -> Generator[Flat, None, None]:
-        """Return another Flat object."""
-        yield Flat(
+    def another_home(self) -> Generator[Home, None, None]:
+        """Return another Home object."""
+        yield Home(
             url="https://another.url",
             address="Another address",
             price=2000,
@@ -38,19 +38,19 @@ class TestStoreToCache:
             year=2001,
         )
 
-    def test_flat_is_stored(self, flat: Flat) -> None:
-        """Test that a flat is stored."""
+    def test_home_is_stored(self, home: Home) -> None:
+        """Test that a home is stored."""
         cache_path = Path(".test_cache")
-        store_to_cache(flats=[flat], email="no-email", cache_path=cache_path)
+        store_to_cache(homes=[home], email="no-email", cache_path=cache_path)
         with cache_path.open() as file:
             assert file.read() == '{"id": "some.url", "email": "no-email"}\n'
         cache_path.unlink()
 
-    def test_multiple_flats_are_stored(self, flat: Flat, another_flat: Flat) -> None:
-        """Test that multiple flats are stored."""
+    def test_multiple_homes_are_stored(self, home: Home, another_home: Home) -> None:
+        """Test that multiple homes are stored."""
         cache_path = Path(".test_cache")
         store_to_cache(
-            flats=[flat, another_flat], email="no-email", cache_path=cache_path
+            homes=[home, another_home], email="no-email", cache_path=cache_path
         )
         with cache_path.open() as file:
             assert (
@@ -60,13 +60,13 @@ class TestStoreToCache:
             )
         cache_path.unlink()
 
-    def test_no_duplicates_are_stored(self, flat: Flat) -> None:
+    def test_no_duplicates_are_stored(self, home: Home) -> None:
         """Test that no duplicates are stored."""
         cache_path = Path(".test_cache")
-        store_to_cache(flats=[flat, flat], email="no-email", cache_path=cache_path)
+        store_to_cache(homes=[home, home], email="no-email", cache_path=cache_path)
         with cache_path.open() as file:
             assert file.read() == '{"id": "some.url", "email": "no-email"}\n'
-        store_to_cache(flats=[flat], email="no-email", cache_path=cache_path)
+        store_to_cache(homes=[home], email="no-email", cache_path=cache_path)
         with cache_path.open() as file:
             assert file.read() == '{"id": "some.url", "email": "no-email"}\n'
         cache_path.unlink()