+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Added the ability to search for different property types, with the new
+  `--property-type` (`-t`) option. The available property types are `ejerlejlighed`,
+  `andelslejlighed` and `house`. You can specify multiple property types with, e.g.,
+  `bolig-ping <other-arguments> -t ejerlejlighed -t house`. It defaults to
+  `ejerlejlighed`, which is the only type that was previously supported.
 - Added new `--min-monthly-fee` and `--max-monthly-fee` options to filter flats by
   monthly fee. The default is no minimum or maximum monthly fee.
 - Added new `--cache/--no-cache` flag, which allows you to disable the cache. The